+++ Assets/DataGateways/WordPressPostDataGateway.class.php
@@ -1,4 +1,7 @@
 <?php
+
+namespace SimpleRSSFeedToPost;
+
 class WordPressPostDataGateway
 {
