+++ Assets/DataStructures/AppDefaultValues.class.php
@@ -1,4 +1,6 @@
 <?php
+namespace SimpleRSSFeedToPost;
+
 function appTitle()
 {
     return 'Simple RSS Feed to Post';
