+++ Assets/DataStructures/RSSDataTemplate.class.php
@@ -1,5 +1,5 @@
 <?php
-
+namespace SimpleRSSFeedToPost;
 class RSSDataTemplate
 {
     const RSSTITLE = 'title';
