+++ Assets/DataStructures/WordPressPostDataTemplate.class.php
@@ -1,5 +1,5 @@
 <?php
-
+namespace SimpleRSSFeedToPost;
 class WordPressPostDataTemplate
 {
     const WORDPRESSPOSTAUTHOR = 'post_author';
