--- Assets/Install/Activation.php
@@ -1,7 +1,0 @@
-<?php
-class Activation
-{
-    public static function activate()
-    {
-    }
-}
