+++ Assets/Settings/Defaults.class.php
