+++ Assets/Settings/SettingsRegistration.class.php
