+++ Assets/Systems/FeedURISystem.class.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace SimpleRSSFeedToPost;
+
+class FeedURISystem
+{
+}
