+++ Assets/Systems/PostingSystem.class.php
@@ -1,5 +1,6 @@
 <?php
 
+namespace SimpleRSSFeedToPost;
 abstract class PostingSystem
 {
   public static function getPostingSystem()
