+++ Run/Loader.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace SimpleRSSFeedToPost;
+
+class Loader
+{
+    public static function loader()
+    {
+        $applicationRoot = '../Assets/';
+        $loaderExtension = '.class';
+        foreach (glob("$applicationRoot*$loaderExtension.php") as $file) {
+            require_once($file);
+        }
+    }
+}
